refactor(client): api 이름명 통일 (action + Data + Repo), update 함수 API화
