+++ client/src/api/InitiateDataRedux.js
@@ -1,8 +1,8 @@
-import FetchContent from "./FetchContent";
+import FetchContentDB from "./FetchContentDB";
 import { InitContent } from "../redux/BlogContent";
 
 const initiateContentData = async (dispatch) => {
-  const newArr = await FetchContent();
+  const newArr = await FetchContentDB();
   dispatch(InitContent(newArr));
 };
 export default initiateContentData;
