+++ client/src/api/UpdateContentDB.js
@@ -1,0 +1,10 @@
+const UpdateContentDB = async (updateArr) => {
+  await fetch("/contents/update", {
+    method: "PATCH",
+    body: JSON.stringify(updateArr),
+    headers: {
+      "Content-Type": "application/json",
+    },
+  });
+};
+export default UpdateContentDB;
